update env when running predict
